ISSUE #19 * Rename "aggregate.py" as "model.py".
