+++ minos/common/model/__init__.py
@@ -6,7 +6,7 @@ This file is part of minos framework.
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
 
-from .aggregate import (
+from .model import (
     BOOLEAN,
     NULL,
     INT,
